+++ src/Ext/Rules/Body.h
@@ -61,6 +61,9 @@ public:
 		Valueable<bool> ForbidParallelAIQueues_Building;
 
 		Valueable<bool> IronCurtain_KeptOnDeploy;
+		Valueable<ColorStruct> ToolTip_Background_Color;
+		Valueable<int> ToolTip_Background_Opacity;
+		Valueable<float> ToolTip_Background_BlurSize;
 
 		ExtData(RulesClass* OwnerObject) : Extension<RulesClass>(OwnerObject)
 			, Storage_TiberiumIndex { -1 }
@@ -93,6 +96,9 @@ public:
 			, ForbidParallelAIQueues_Navy { false }
 			, ForbidParallelAIQueues_Vehicle { false }
 			, IronCurtain_KeptOnDeploy { true }
+			, ToolTip_Background_Color { { 0, 0, 0 } }
+			, ToolTip_Background_Opacity { 100 }
+			, ToolTip_Background_BlurSize { 0.0f }
 		{ }
 
 		virtual ~ExtData() = default;